refactor: drop all api files down a directory
