+++ .gitignore
@@ -16,3 +16,4 @@ __pycache__
 .DS_Store
 .vscode
 volumes
+devcontainer.json
