+++ requirements.in
@@ -1,0 +1,9 @@
+django
+djangorestframework
+psycopg2-binary
+sqlparse
+asgiref
+pymilvus==2.3.6
+sentence-transformers
+cffi
+retry
