+++ requirements.txt
@@ -1,9 +1,145 @@
-django
-djangorestframework
-psycopg2-binary
-sqlparse
-asgiref
+# This file was autogenerated by uv via the following command:
+#    uv pip compile requirements.in
+argon2-cffi==23.1.0
+    # via minio
+argon2-cffi-bindings==21.2.0
+    # via argon2-cffi
+asgiref==3.7.2
+    # via django
+certifi==2024.2.2
+    # via
+    #   minio
+    #   requests
+cffi==1.16.0
+    # via argon2-cffi-bindings
+charset-normalizer==3.3.2
+    # via requests
+decorator==5.1.1
+    # via retry
+django==5.0.3
+    # via djangorestframework
+djangorestframework==3.15.0
+environs==9.5.0
+    # via pymilvus
+filelock==3.13.1
+    # via
+    #   huggingface-hub
+    #   torch
+    #   transformers
+fsspec==2024.3.1
+    # via
+    #   huggingface-hub
+    #   torch
+grpcio==1.60.0
+    # via pymilvus
+huggingface-hub==0.21.4
+    # via
+    #   sentence-transformers
+    #   tokenizers
+    #   transformers
+idna==3.6
+    # via requests
+jinja2==3.1.3
+    # via torch
+joblib==1.3.2
+    # via scikit-learn
+markupsafe==2.1.5
+    # via jinja2
+marshmallow==3.21.1
+    # via environs
+minio==7.2.5
+    # via pymilvus
+mpmath==1.3.0
+    # via sympy
+networkx==3.2.1
+    # via torch
+numpy==1.26.4
+    # via
+    #   pandas
+    #   pyarrow
+    #   scikit-learn
+    #   scipy
+    #   sentence-transformers
+    #   transformers
+packaging==24.0
+    # via
+    #   huggingface-hub
+    #   marshmallow
+    #   transformers
+pandas==2.2.1
+    # via pymilvus
+pillow==10.2.0
+    # via sentence-transformers
+protobuf==5.26.0
+    # via pymilvus
+psycopg2-binary==2.9.9
+py==1.11.0
+    # via retry
+pyarrow==15.0.2
+    # via pymilvus
+pycparser==2.21
+    # via cffi
+pycryptodome==3.20.0
+    # via minio
 pymilvus==2.3.6
-sentence-transformers
-cffi
-retry
+python-dateutil==2.9.0.post0
+    # via pandas
+python-dotenv==1.0.1
+    # via environs
+pytz==2024.1
+    # via pandas
+pyyaml==6.0.1
+    # via
+    #   huggingface-hub
+    #   transformers
+regex==2023.12.25
+    # via transformers
+requests==2.31.0
+    # via
+    #   huggingface-hub
+    #   pymilvus
+    #   transformers
+retry==0.9.2
+safetensors==0.4.2
+    # via transformers
+scikit-learn==1.4.1.post1
+    # via sentence-transformers
+scipy==1.12.0
+    # via
+    #   scikit-learn
+    #   sentence-transformers
+sentence-transformers==2.5.1
+setuptools==69.2.0
+    # via pymilvus
+six==1.16.0
+    # via python-dateutil
+sqlparse==0.4.4
+    # via django
+sympy==1.12
+    # via torch
+threadpoolctl==3.3.0
+    # via scikit-learn
+tokenizers==0.15.2
+    # via transformers
+torch==2.2.1
+    # via sentence-transformers
+tqdm==4.66.2
+    # via
+    #   huggingface-hub
+    #   sentence-transformers
+    #   transformers
+transformers==4.38.2
+    # via sentence-transformers
+typing-extensions==4.10.0
+    # via
+    #   huggingface-hub
+    #   minio
+    #   torch
+tzdata==2024.1
+    # via pandas
+ujson==5.9.0
+    # via pymilvus
+urllib3==2.2.1
+    # via
+    #   minio
+    #   requests
