+++ rfp_api/views.py
@@ -1,19 +1,18 @@
-from django.http import HttpResponse
 from django.shortcuts import redirect, render
 
-from ..models import Answer, Question
+from .models import Answer, Question
 
 
-def indexPageView(request):
+def index_page_view(request):
     return render(request, "index.html")
 
 
-def deleteQuestionPageView(request, iQuestionID):
+def delete_question_page_view(request, iQuestionID):
     Question.objects.get(id=iQuestionID).delete()
-    return redirect(listQuestionsPageView)
+    return redirect(list_questions_page_view)
 
 
-def editQuestionPageView(request, iQuestionID):
+def edit_question_page_view(request, iQuestionID):
     question = Question.objects.get(id=iQuestionID)
 
     if request.method == "POST":
@@ -21,19 +20,19 @@ def editQuestionPageView(request, iQuestionID):
         question.answer = request.POST.get("answer")
         question.save()
 
-        return redirect(listQuestionsPageView)
+        return redirect(list_questions_page_view)
 
     context = {"question": question}
     return render(request, "editQuestion.html", context)
 
 
-def listAnswersPageView(request):
+def list_answers_page_view(request):
     data = Answer.objects.all()
     context = {"answers": data}
     return render(request, "answerList.html", context)
 
 
-def listQuestionsPageView(request):
+def list_questions_page_view(request):
     if request.method == "POST":
         new_question = Question()
         new_question.text = request.POST.get("question")
